pyproject.toml
[tool.poetry]
name = "imsim"
version = "0.1.0"
description = ""
authors = ["Carlos Miranda <carlos.miralop@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.38.2"
faiss-gpu = "^1.7.2"
torch = "^2.2.1"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
pre-commit = "^3.6.2"

[tool.poetry.group.cli.dependencies]
typer = "^0.9.0"

[tool.poetry.scripts]
imsim-cli = "imsim.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
